fix: ignore remove view if project disposed

+++ plugin/src/main/kotlin/spp/jetbrains/sourcemarker/view/window/LiveViewChartWindow.kt
@@ -124,6 +124,9 @@ class LiveViewChartWindow(
         consumer?.unregister()
         consumer = null
         liveView.subscriptionId?.let {
+            if (project.isDisposed) {
+                return //no need to remove view
+            }
             viewService.removeLiveView(it).onFailure {
                 log.error("Failed to pause live view", it)
             }
